refactor(health): delete internal/k8s/health.go, re-point all callers (no shims)
Completes the Go consolidation: pkg/health is now the ONLY backend health
classifier. internal/k8s/health.go — the shim layer added earlier to avoid
touching callers — is deleted entirely; every caller (detect, detect_node,
dashboard, mcp, applications) now calls pkg/health directly.

- ClassifyPodHealth/PodProblemReason/PodProblemMessage/PodRestartContext/
  IsPodUnschedulable/PodCrashLoopDiagnosis/PodHasReadinessProbeFailure: callers
  re-pointed to health.* (string consumers use Verdict.LegacyString()).
- ClassifyNodeHealth + NodeHealth moved to pkg/health.Node / health.NodeHealth.
- Detector-only reason constants moved to internal/k8s/reasons.go.
- Test coverage for the recovery guards, crashloop diagnosis ordering, and node
  classification ported into pkg/health (recovery_test.go, node_test.go).

No shims, no duplicate classifiers, one rank ordering. Both modules build, vet,
and pass all tests.
